Refactor similarity functions
