--- src/NLP/GingerbreadAI.NLP.Word2Vec/Properties/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("GingerbreadAI.NLP.Word2Vec.Test")]
