+++ src/NLP/GingerbreadAI.NLP.Word2Vec/SimilarityFunctions/SimilarityFunctionType.cs
@@ -1,0 +1,7 @@
+﻿namespace GingerbreadAI.NLP.Word2Vec.SimilarityFunctions
+{
+    public enum SimilarityFunctionType
+    {
+        Cosine = 1,
+    }
+}
